Repackage src/test/classes to avoid split package with src/main/java
