+++ src/test/java/io/vertx/tests/ProxyClientPipelinedTest.java
@@ -8,23 +8,18 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  */
-package io.vertx.httpproxy;
+package io.vertx.tests;
 
-import io.vertx.ext.unit.TestContext;
+import io.vertx.httpproxy.ProxyOptions;
 
 /**
  * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
  */
-public class ProxyClientNotPersistentTest extends ProxyClientKeepAliveTest {
+public class ProxyClientPipelinedTest extends ProxyClientKeepAliveTest {
 
-  public ProxyClientNotPersistentTest(ProxyOptions options) {
+  public ProxyClientPipelinedTest(ProxyOptions options) {
     super(options);
-    keepAlive = false;
-    pipelining = false;
-  }
-
-  public void testChunkedTransferEncodingRequest(TestContext ctx) {
-    // super.testChunkedTransferEncodingRequest(ctx);
-    // Does not pass for now - only when run in single
+    keepAlive = true;
+    pipelining = true;
   }
 }
