+++ src/test/java/io/vertx/tests/cache/CacheMaxAgeTest.java
@@ -8,7 +8,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  */
-package io.vertx.httpproxy.cache;
+package io.vertx.tests.cache;
 
 import io.vertx.core.MultiMap;
 import io.vertx.core.http.HttpHeaders;
