+++ src/test/java/io/vertx/tests/parsing/ParseTest.java
@@ -8,7 +8,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  */
-package io.vertx.httpproxy.parsing;
+package io.vertx.tests.parsing;
 
 import io.vertx.httpproxy.impl.CacheControl;
 import org.junit.Assert;
